+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * 【core】          ImageUtil.pressText增加抗锯齿（pr#19@Gitee）
 * 【core】          BeanUtil.getPropertyDescriptors去除class属性（issue#IKVKR@Gitee）
 * 【json】           putByPath方法针对空的规则变更（issue#IKX2H@Gitee）
+* 【captcha】     增加CodeGenerator，可自定义验证码文字生成策略（issue#IL3YH@Gitee）
 
 ### Bug修复
 * 【core】          修复JSONUtil.formatJsonStr引号换行问题（issue#IKMMK@Gitee）
@@ -20,6 +21,7 @@
 * 【json】          修复当Bean为私有类时无法实例化导致的JSON转换问题（感谢@【上海】风景）
 * 【json】          修复Bean中有Object字段时toBean产生的问题（感谢@【上海】风景）
 * 【core】          修复XmlUtil关闭XXE避免XXE攻击
+* 【poi】            修复Excel03SaxReader读取小数的问题（感谢@【深圳】rm -rf /）
 
 -------------------------------------------------------------------------------------------------------------